Migration to Cleanbasket Repository

+++ CleanBasket/Keychain.m
@@ -50,7 +50,7 @@
     
     OSStatus status = SecItemAdd((__bridge CFDictionaryRef)dict, NULL);
     if(errSecSuccess != status) {
-        NSLog(@"Unable add item with key =%@ error:%ld",key,status);
+        NSLog(@"Unable add item with key =%@ error:%d",key,(int)status);
     }
     return (errSecSuccess == status);
 }
@@ -63,7 +63,7 @@
     OSStatus status = SecItemCopyMatching((__bridge CFDictionaryRef)dict,&result);
     
     if( status != errSecSuccess) {
-        NSLog(@"Unable to fetch item for key %@ with error:%ld",key,status);
+        NSLog(@"Unable to fetch item for key %@ with error:%d",key,(int)status);
         return nil;
     }
     
@@ -79,7 +79,7 @@
     
     OSStatus status = SecItemUpdate((__bridge CFDictionaryRef)dictKey, (__bridge CFDictionaryRef)dictUpdate);
     if(errSecSuccess != status) {
-        NSLog(@"Unable add update with key =%@ error:%ld",key,status);
+        NSLog(@"Unable add update with key =%@ error:%d",key,(int)status);
     }
     return (errSecSuccess == status);
     
@@ -91,7 +91,7 @@
     NSMutableDictionary *dict = [self prepareDict:key];
     OSStatus status = SecItemDelete((__bridge CFDictionaryRef)dict);
     if( status != errSecSuccess) {
-        NSLog(@"Unable to remove item for key %@ with error:%ld",key,status);
+        NSLog(@"Unable to remove item for key %@ with error:%d",key,(int)status);
         return NO;
     }
     return  YES;
